Fixed report status to u8

+++ src/reporter/report.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 /// Represent a response from a executed command
 #[derive(Serialize, Deserialize, Clone, Debug)]
 pub struct Report {
-    pub status: i32,
+    pub status: u8,
     pub stdout: Vec<u8>,
     pub stderr: Vec<u8>,
 }
